Move utils.tsx into form directory
